Namespace library with github name.
While this is header-only, and users are expected
to download the header and commit it to their project, it's better to
have unique namespace for the header by default.

This has two reasons:

  1. I wish to enable installing it with conan.io, and want to reduce
     the possibility for namespace clashes. The term "subprocess" is
     pretty overloaded.
  2. Currently the README and the CMakeLists differ in their include
     path, it's better to be consistent, even though users can put it
     wherever they please in their project.

Since most users are using it as header only anyway, I don't think it'd
make much diference.